+++ lib/src/flutter/utilities.dart
@@ -3,3 +3,5 @@ export 'utilities/padding.dart';
 export 'utilities/border_radius.dart';
 
 export 'utilities/colors.dart';
+
+export 'utilities/responsive.dart';